Working on Movie Dialog

+++ src/pmcollection/be/Movie.java
@@ -6,12 +6,12 @@ import java.util.List;
 public class Movie {
     private int id;
     private String name;
-    private double rating;
+    private float rating;
     private String filelink;
     private LocalDate lastview;
     private List<Category> categories;
 
-    public Movie(int id, String name, double rating, String filelink, LocalDate lastview) {
+    public Movie(int id, String name, float rating, String filelink, LocalDate lastview) {
         this.id = id;
         this.name = name;
         this.rating = rating;
@@ -31,7 +31,7 @@ public class Movie {
         this.name = name;
     }
 
-    public double getRating() {
+    public float getRating() {
         return rating;
     }
